+++ CHANGELOG.md
@@ -9,6 +9,12 @@ Versioning].
 
 ## [Unreleased] <!-- #release:date -->
 
+* Add support for non-default status filters in `Client::list_invoices`.
+
+* Return `hosted_invoice_url` as part of `Invoice`, when available.
+
+* Fix bug causing invoices without `invoice_pdf` set to cause errors.
+
 ## [0.3.0] - 2023-04-13
 
 * Add support for request idempotency keys to `Client::create_customer` and